test: run KMP test suites on JS by moving them to commonTest
Moves the platform-agnostic jvmTest suites of f2-dsl-function (10 files),
f2-dsl-cqrs (3 of 4) and f2-client-ktor-common (1) to commonTest, rewriting
JUnit 5 + AssertJ to kotlin.test and suspend test functions to runTest so they
compile for both targets. Adds a commonTest smoke test for the
HttpClientBuilder expect/actual pair, which covers the previously uncovered
jsMain actual.

Also fixes the make target: `./gradlew test` registers no task at all for KMP
modules, so `make test` was running zero tests for f2-dsl and f2-client. It now
runs `allTests test`.

Closes #125
