+++ infra/make/libs.mk
@@ -11,8 +11,10 @@ lint:
 build:
 	VERSION=$(VERSION) ./gradlew clean build publishToMavenLocal -x test
 
+# `test` is a JVM-only Gradle task: Kotlin Multiplatform modules (f2-dsl, f2-client) do not
+# register one, their tests live behind `allTests` (jvmTest + jsTest).
 test:
-	./gradlew test
+	./gradlew allTests test
 
 check:
 	VERSION=$(VERSION) ./gradlew sonar
